+++ packages/orq-rc/src/orq_ai_sdk/models/createconversationop.py
@@ -193,7 +193,7 @@ class CreateConversationResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kergt7zemjn0dy6m47k8xkym"
+        "conv_01kerte095zg6ndtmc3kfppmcm"
     )
     r"""Unique conversation identifier with `conv_` prefix."""