+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -239,7 +239,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KERGT8D3Q4867X0JCCQ7JGW7"
+        "01KERTE115F6NNJEG7CPNGGMRG"
     )
     r"""The unique identifier of the data source"""