+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KERGT8D2ZWJV05A4T5S781ZR"
+        "01KERTE114Y0A53GDXC8QCQKGH"
     )
     r"""The unique identifier of the data source"""