+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -121,5 +121,5 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-01-12T07:12:53.135Z")
+    updated: Optional[datetime] = parse_datetime("2026-01-12T10:00:55.882Z")
     r"""The date and time the resource was last updated"""