+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KERGT8D7371S66DPB318RRBQ"
+        "01KERTE117X31ZWNFXWT70CS10"
     )
     r"""The unique identifier of the data source"""