+++ packages/orq-rc/RELEASES.md
@@ -4589,3 +4589,13 @@ Based on:
 - [python v4.2.0-rc.21] packages/orq-rc
 ### Releases
 - [PyPI v4.2.0-rc.21] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.21 - packages/orq-rc
+
+## 2026-01-12 10:03:47
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.684.0 (2.793.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.2.0-rc.22] packages/orq-rc
+### Releases
+- [PyPI v4.2.0-rc.22] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.22 - packages/orq-rc